+++ src/creators/at-rule.js
@@ -1,7 +1,0 @@
-const createAtRule = (name, hasBlock) => ({
-  name,
-  hasBlock,
-  type: 'at-rule',
-});
-
-export default createAtRule;