+++ src/creators/logical-group.js
@@ -1,9 +1,0 @@
-const createLogicalGroup = (groupName, properties) => ({
-  groupName,
-  properties,
-  emptyLineBefore: 'always',
-  noEmptyLineBetween: true,
-  order: 'flexible',
-});
-
-export default createLogicalGroup;