+++ src/creators/rule.js
@@ -1,6 +1,0 @@
-const createRule = (selector) => ({
-  selector: `&${selector}`,
-  type: 'rule',
-});
-
-export default createRule;