+++ src/creators/rule.ts
@@ -1,0 +1,13 @@
+/**
+ * The function `createRule` in TypeScript creates
+ * a CSS rule with a specified selector.
+ * @param {string} selector - The `selector` parameter in the `createRule`
+ * function is a string that
+ * represents the CSS selector for which a rule is being created.
+ */
+const createRule = (selector: string) => ({
+  selector: `&${selector}`,
+  type: 'rule',
+});
+
+export default createRule;